Move stuff from import actions into import_mixins.
